Add reporting of sub trees ordered by activation scores of their root nodes
This is not the most intutitve output, but it can be postprocessed if needed.

Search for subtree-activations in the output file.

+++ src/tools/nodestats/SubTree.java
@@ -3,9 +3,9 @@ package tools.nodestats;
 final class SubTree {
   private final AstNode rootNode;
 
-  final int score;
+  final long score;
 
-  SubTree(final AstNode rootNode, final int score) {
+  SubTree(final AstNode rootNode, final long score) {
     this.rootNode = rootNode;
     this.score = score;
   }
@@ -14,7 +14,7 @@ final class SubTree {
     return rootNode;
   }
 
-  public int getScore() {
+  public long getScore() {
     return score;
   }
